scripts/requirements.txt
requests>=2.32.3
pypdf>=4.3.1
langchain-text-splitters>=0.2.3
langchain-core>=0.2.5
tqdm>=4.66.5
python-dateutil>=2.9.0.post0
python-slugify>=8.0.4
